Major refactor for cleaner project structure
