+++ src/main/ipc-events/get-twitch-token.js
@@ -1,0 +1,6 @@
+import { ipcMain } from 'electron';
+import { getToken } from '../twitch/auth';
+
+ipcMain.handle('twitch::auth::get-token', () => {
+  const token = getToken();
+});
