+++ src/main/twitch/auth/index.js
@@ -1,0 +1,5 @@
+import { getToken } from 'get-token';
+
+export default {
+  getToken,
+};
