+++ src/renderer/App.vue
