+++ src/renderer/assets/logo.png
