+++ src/renderer/public/favicon.ico
