+++ src/renderer/views/Home.vue
