--- src/twitch/auth/index.js
@@ -1,5 +1,0 @@
-import tokenResponse from 'get-token';
-
-export default {
-  tokenResponse,
-};
